requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
fairscale==0.4.4
filelock==3.7.0
fonttools==4.33.3
huggingface-hub==0.6.0
idna==3.3
imageio==2.19.2
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
pycocoevalcap==1.2
pycocotools==2.0.4
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
sacremoses==0.0.53
six==1.16.0
timm==0.4.12
tokenizers==0.10.3
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.15.0
typing-extensions==4.2.0
urllib3==1.26.9